submission_BinaryAgents/requirements.txt
google-genai
python-dotenv
pydantic
markdown
google-api-python-client
google-auth
